Fixed out of map crash and zombie AI
Zombie now features a breathtaking +5 IQ and slightly spring-ey-er legs.
Jumps in both directions, and too much.
Game no longer crashes when an entity falls out of the map.
Added a function to automatically index quads on a tilesheet.
Tileset is now 500% more shilled.
Camera can now be scaled correctly when zoomed in.